fix: address review findings on PR #60
- test_visualizer.py: fix prev_db dtype float64→float32 to match
  production (MEDIUM)
- visualizer.py: clear _last_broadcast when clients set empties,
  so reconnecting clients get current frame immediately (LOW)

+++ tests/test_visualizer.py
@@ -67,7 +67,7 @@ class TestAnalyzePcm:
 
     def setup_method(self):
         """Reset global state before each test."""
-        visualizer.prev_db = np.full(visualizer.NUM_BANDS, visualizer.NOISE_FLOOR, dtype=np.float64)
+        visualizer.prev_db = np.full(visualizer.NUM_BANDS, visualizer.NOISE_FLOOR, dtype=np.float32)
         visualizer.audio_ring = np.zeros(visualizer.FFT_SIZE, dtype=np.float32)
 
     def test_silence_returns_noise_floor(self):
